server/requirements.txt
asgiref==3.2.3
boto3==1.13.13
botocore==1.16.13
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
jmespath==0.10.0
Pillow==7.1.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
